update useStrapi for v5
